+++ src/index.ts
@@ -1,5 +1,5 @@
 import ReadECG from "./utils/ReadECG"; //Development
-import { SOP_CLASS_UIDS, WAVE_FORM_BITS_STORED } from "./constants/Constants";
+import { SOP_CLASS_UIDS, WAVE_FORM_BITS_STORED, KEY_UNIT_INFO } from "./constants/Constants";
 import DicomECGViewer from "./viewer/DicomECGViewer";
 
-export { ReadECG, DicomECGViewer, SOP_CLASS_UIDS, WAVE_FORM_BITS_STORED };
+export { ReadECG, DicomECGViewer, SOP_CLASS_UIDS, WAVE_FORM_BITS_STORED, KEY_UNIT_INFO };